+++ src/main/java/com/codenear/butterfly/member/application/MemberService.java
@@ -2,16 +2,10 @@ package com.codenear.butterfly.member.application;
 
 import static com.codenear.butterfly.s3.domain.S3Directory.PROFILE_IMAGE;
 
-import com.codenear.butterfly.auth.annotation.PasswordValidator;
-import com.codenear.butterfly.auth.domain.dto.AuthRegisterDTO;
-import com.codenear.butterfly.certify.application.CertifyService;
-import com.codenear.butterfly.certify.domain.CertifyType;
-import com.codenear.butterfly.certify.domain.dto.CertifyRequest;
 import com.codenear.butterfly.global.exception.ErrorCode;
 import com.codenear.butterfly.member.domain.Member;
-import com.codenear.butterfly.member.domain.dto.*;
-import com.codenear.butterfly.auth.annotation.ValidPassword;
-import com.codenear.butterfly.member.domain.enums.VerificationType;
+import com.codenear.butterfly.member.domain.dto.MemberDTO;
+import com.codenear.butterfly.member.domain.dto.MemberInfoDTO;
 import com.codenear.butterfly.member.domain.repository.member.MemberRepository;
 import com.codenear.butterfly.member.exception.MemberException;
 import com.codenear.butterfly.notify.alarm.infrastructure.AlarmRepository;
@@ -20,7 +14,6 @@ import com.codenear.butterfly.s3.application.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,8 +26,6 @@ public class MemberService {
     private final PointService pointService;
     private final S3Service s3Service;
     private final AlarmRepository alarmRepository;
-    private final CertifyService certifyService;
-    private final PasswordEncoder passwordEncoder;
 
     public MemberInfoDTO getMemberInfo(MemberDTO memberDTO) {
         Integer pointValue = pointService.loadPointByMemberId(memberDTO.getId()).getPoint();
@@ -80,79 +71,4 @@ public class MemberService {
         return memberRepository.findById(memberId)
                 .orElseThrow(() -> new MemberException(ErrorCode.SERVER_ERROR, null));
     }
-
-    @Transactional(readOnly = true)
-    private void validateMemberExistsByPhone(String phoneNumber) {
-        if (!memberRepository.findByPhoneNumber(phoneNumber).isPresent()) {
-            throw new MemberException(ErrorCode.MEMBER_NOT_FOUND_BY_PHONE, null);
-        }
-    }
-
-    public void sendFindEmailcode(String phoneNumber) {
-        validateMemberExistsByPhone(phoneNumber);
-        certifyService.sendCertifyCode(phoneNumber, CertifyType.CERTIFY_PHONE);
-    }
-
-    @Transactional(readOnly = true)
-    private Member loadMemberByPhoneNumber(String phoneNumber) {
-        return memberRepository.findByPhoneNumber(phoneNumber)
-                .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND_BY_PHONE, null));
-    }
-
-    public String findEmail(CertifyRequest request) {
-        certifyService.checkCertifyCode(request, CertifyType.CERTIFY_PHONE);
-
-        Member member = loadMemberByPhoneNumber(request.phoneNumber());
-        return member.getEmail();
-    }
-
-    @Transactional(readOnly = true)
-    private void validateMemberExistsByEmail(String email) {
-        if (!memberRepository.findByEmail(email).isPresent()) {
-            throw new MemberException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL, null);
-        }
-    }
-
-    public void sendFindPasswordCode(FindPasswordRequestDTO request){
-        switch (request.getType()){
-            case PHONE -> {
-                validateMemberExistsByPhone(request.getIdentifier());
-                certifyService.sendCertifyCode(request.getIdentifier(), CertifyType.CERTIFY_PHONE);
-            }
-            case EMAIL -> {
-
-            }
-        }
-    }
-
-    public void verifyFindPasswordCode(VerifyFindPasswordRequestDTO request){
-        switch (request.getType()){
-            case PHONE -> {
-                CertifyRequest certifyRequest = new CertifyRequest(
-                        request.getIdentifier(),
-                        request.getCertifyCode()
-                );
-                certifyService.checkCertifyCode(certifyRequest, CertifyType.CERTIFY_PHONE);
-            }
-            case EMAIL -> {
-
-            }
-        }
-    }
-
-    @CacheEvict(value = "userCache, memberCache", key = "#memberId")
-    public void resetPassword(ResetPasswordRequestDTO request) {
-
-        if (request.getType() == VerificationType.PHONE) {
-            Member member = loadMemberByPhoneNumber(request.getIdentifier());
-
-            String encodedPassword = passwordEncoder.encode(request.getNewPassword());
-            member.updatePassword(encodedPassword);
-
-            memberRepository.save(member);
-
-        } else if (request.getType() == VerificationType.EMAIL) {
-        }
-    }
-
 }
